fix MSVC level 4 warnings
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/535



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@2259 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/i18n/rbt_set.cpp
@@ -217,7 +217,7 @@ TransliterationRuleSet::findMatch(const Replaceable& text,
     /* We only need to check our indexed bin of the rule table,
      * based on the low byte of the first key character.
      */
-    int16_t x = text.charAt(pos.start) & 0xFF;
+    int16_t x = (int16_t) (text.charAt(pos.start) & 0xFF);
     for (int32_t i=index[x]; i<index[x+1]; ++i) {
         if (rules[i]->matches(text, pos, data, filter)) {
             return rules[i];
@@ -264,7 +264,7 @@ TransliterationRuleSet::findIncrementalMatch(const Replaceable& text,
      * based on the low byte of the first key character.
      */
     isPartial = FALSE;
-    int16_t x = text.charAt(pos.start) & 0xFF;
+    int16_t x = (int16_t) (text.charAt(pos.start) & 0xFF);
     for (int32_t i=index[x]; i<index[x+1]; ++i) {
         int32_t match = rules[i]->getMatchDegree(text, pos, data, filter);
         switch (match) {
